Add chunked fetching of deltas for very long change sets, add some more amd64 safety, and fix a dumb login overoptimisation
svn path=/branches/KDE/3.5/kdepim/; revision=612498

+++ kresources/groupwise/kabc_resourcegroupwise.h
@@ -158,7 +158,7 @@ class KDE_EXPORT ResourceGroupwise : public ResourceCached
     KPIM::ProgressItem *mUABProgress;
     QString mJobData;
     ResourceState mState;
-    int mServerFirstSequence, mServerLastSequence, mServerLastPORebuildTime;
+    unsigned long mServerFirstSequence, mServerLastSequence, mServerLastPORebuildTime;
 
     bool mLimitedMode;
 };
